+++ library/Solarium/QueryType/Select/RequestBuilder/Component/DisMax.php
@@ -68,7 +68,15 @@ class DisMax implements ComponentRequestBuilderInterface
         $request->addParam('ps', $component->getPhraseSlop());
         $request->addParam('qs', $component->getQueryPhraseSlop());
         $request->addParam('tie', $component->getTie());
-        $request->addParam('bq', $component->getBoostQuery());
+
+        // add boostqueries to request
+        $boostQueries = $component->getBoostQueries();
+        if (count($boostQueries) !== 0) {
+            foreach ($boostQueries as $boostQuery) {
+                $request->addParam('bq', $boostQuery->getQuery());
+            }
+        }
+
         $request->addParam('bf', $component->getBoostFunctions());
 
         return $request;